Move protocol out of utils
